Handle new opcodes / interrupts
Implement support for all of the new opcodes.
Hardware right now is hard-coded to 0 devices, as we have no information
on how hardware itself will work.

Add a method TriggerSoftwareInterrupt() that will cause the interrupt to
be processed after the currently-executing instruction is complete. If
multiple interrupts are fired simultaneously, they will all be
processed. Yes, interrupts can interrupt other interrupts, but there's
no alternative with the current spec.