Update for spec v1.4
Temporarily disable hardware-triggered interrupts. These will need to be
asynchronous (e.g. the interrupt may not be processed for a few cycles,
and the hardware needs to be notified when it is), which requires
tighter integration with the hardware APIs. Since the hardware APIs
don't actually exist yet, we can't process the interrupts.